chore(translator): move non top-level methods to separate files
